requirements.txt
h5py
torch-scatter
open3d
tensorboard
meshio
gdown
pyyaml